tensorflow_toolkit/text_detection/requirements.txt
absl-py==0.7.1
astor==0.7.1
gast==0.2.2
google-pasta==0.1.5
grpcio==1.19.0
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
numpy==1.16.2
opencv-python==4.1.0.25
Polygon3==3.0.8
protobuf==3.7.1
PyYAML==5.1
six==1.12.0
tb-nightly==1.14.0a20190301
tensorflow-gpu==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
tqdm==4.31.1
Werkzeug==0.15.2
